Added fixes from feedback
