+++ src/main/scala/ch/epfl/bluebrain/nexus/kg/Main.scala
@@ -21,10 +21,11 @@ import ch.epfl.bluebrain.nexus.kg.config.AppConfig.{ElasticConfig, SparqlConfig}
 import ch.epfl.bluebrain.nexus.kg.config.Settings
 import ch.epfl.bluebrain.nexus.kg.indexing.Indexing
 import ch.epfl.bluebrain.nexus.kg.persistence.TaskAggregate
+import ch.epfl.bluebrain.nexus.kg.resolve.ProjectResolution
 import ch.epfl.bluebrain.nexus.kg.resources.Repo.Agg
 import ch.epfl.bluebrain.nexus.kg.resources.attachment.AttachmentStore
 import ch.epfl.bluebrain.nexus.kg.resources.attachment.AttachmentStore.{AkkaIn, AkkaOut}
-import ch.epfl.bluebrain.nexus.kg.resources.{Repo, Resources, ResourcesImpl}
+import ch.epfl.bluebrain.nexus.kg.resources.{Repo, Resources}
 import ch.epfl.bluebrain.nexus.kg.routes.{Clients, ResourceRoutes, ServiceDescriptionRoutes}
 import ch.epfl.bluebrain.nexus.service.http.directives.PrefixDirectives._
 import ch.epfl.bluebrain.nexus.sourcing.akka.{ShardingAggregate, SourcingAkkaSettings}
@@ -92,8 +93,8 @@ object Main {
     implicit val store             = new AttachmentStore[Task, AkkaIn, AkkaOut]
     implicit val indexers          = clients
     implicit val projects          = Projects.task()
-    implicit val elastic           = clients.elastic
-    val resources: Resources[Task] = new ResourcesImpl[Task]()
+    implicit val projectResolution = ProjectResolution.task(projects)
+    val resources: Resources[Task] = Resources[Task]
     val resourceRoutes             = ResourceRoutes(resources).routes
     val apiRoutes                  = uriPrefix(appConfig.http.publicUri)(resourceRoutes)
     val serviceDesc                = ServiceDescriptionRoutes(appConfig.description).routes
