--- src/main/scala/ch/epfl/bluebrain/nexus/kg/resources/ResourcesImpl.scala
@@ -1,442 +1,0 @@
-package ch.epfl.bluebrain.nexus.kg.resources
-
-import java.util.UUID
-
-import cats.Monad
-import cats.data.{EitherT, OptionT}
-import cats.syntax.flatMap._
-import ch.epfl.bluebrain.nexus.commons.es.client.ElasticClient
-import ch.epfl.bluebrain.nexus.commons.http.HttpClient
-import ch.epfl.bluebrain.nexus.commons.types.search.QueryResults.UnscoredQueryResults
-import ch.epfl.bluebrain.nexus.commons.types.search.{Pagination, QueryResults}
-import ch.epfl.bluebrain.nexus.iam.client.types.Identity
-import ch.epfl.bluebrain.nexus.kg.async.Projects
-import ch.epfl.bluebrain.nexus.kg.config.AppConfig.ElasticConfig
-import ch.epfl.bluebrain.nexus.kg.config.Contexts
-import ch.epfl.bluebrain.nexus.kg.config.Schemas._
-import ch.epfl.bluebrain.nexus.kg.config.Vocabulary._
-import ch.epfl.bluebrain.nexus.kg.indexing.ElasticIndexer
-import ch.epfl.bluebrain.nexus.kg.indexing.View.ElasticView
-import ch.epfl.bluebrain.nexus.kg.resources.Rejection._
-import ch.epfl.bluebrain.nexus.kg.resources.ResourceF.Value
-import ch.epfl.bluebrain.nexus.kg.resources.ResourcesImpl.SchemaContext
-import ch.epfl.bluebrain.nexus.kg.resources.attachment.Attachment.{BinaryAttributes, BinaryDescription}
-import ch.epfl.bluebrain.nexus.kg.resources.attachment.AttachmentStore
-import ch.epfl.bluebrain.nexus.kg.resources.syntax._
-import ch.epfl.bluebrain.nexus.kg.search.QueryBuilder
-import ch.epfl.bluebrain.nexus.kg.validation.Validator
-import ch.epfl.bluebrain.nexus.rdf.Graph._
-import ch.epfl.bluebrain.nexus.rdf.Iri.AbsoluteIri
-import ch.epfl.bluebrain.nexus.rdf.Node.{BNode, IriNode}
-import ch.epfl.bluebrain.nexus.rdf.Vocabulary._
-import ch.epfl.bluebrain.nexus.rdf.syntax.circe._
-import ch.epfl.bluebrain.nexus.rdf.syntax.circe.context._
-import ch.epfl.bluebrain.nexus.rdf.syntax.nexus._
-import ch.epfl.bluebrain.nexus.rdf.syntax.node._
-import ch.epfl.bluebrain.nexus.rdf.syntax.node.encoder._
-import ch.epfl.bluebrain.nexus.rdf.{Graph, Iri}
-import io.circe.Json
-
-/**
-  * Resource operations.
-  */
-class ResourcesImpl[F[_]](implicit F: Monad[F],
-                          repo: Repo[F],
-                          projects: Projects[F],
-                          elastic: ElasticClient[F],
-                          esConfig: ElasticConfig)
-    extends Resources[F] { self =>
-
-  def create(projectRef: ProjectRef, base: AbsoluteIri, schema: Ref, source: Json)(
-      implicit identity: Identity): RejOrResource =
-    // format: off
-    for {
-      rawValue       <- materialize(projectRef, source)
-      value          <- checkOrAssignId(Right((projectRef, base)), rawValue)
-      (id, assigned)  = value
-      resource       <- create(id, schema, assigned)
-    } yield resource
-  // format: on
-
-  def create(id: ResId, schema: Ref, source: Json)(implicit identity: Identity): RejOrResource =
-    for {
-      assigned <- materialize(id, source)
-      resource <- create(id, schema, assigned)
-    } yield resource
-
-  /**
-    * Creates a new resource.
-    *
-    * @param id     the id of the resource
-    * @param schema a schema reference that constrains the resource
-    * @param value  the resource representation in json-ld and graph formats
-    */
-  def create(id: ResId, schema: Ref, value: ResourceF.Value)(implicit identity: Identity): RejOrResource = {
-
-    def checkAndJoinTypes(types: Set[AbsoluteIri]): EitherT[F, Rejection, Set[AbsoluteIri]] =
-      EitherT.fromEither(schema.iri match {
-        case `shaclSchemaUri` if types.isEmpty || types.contains(nxv.Schema)      => Right(types + nxv.Schema)
-        case `shaclSchemaUri`                                                     => Left(IncorrectTypes(id.ref, types))
-        case `ontologySchemaUri` if types.isEmpty || types.contains(nxv.Ontology) => Right(types + nxv.Ontology)
-        case `ontologySchemaUri`                                                  => Left(IncorrectTypes(id.ref, types))
-        case _                                                                    => Right(types)
-      })
-
-    //TODO: For now the schema is not validated against the shacl schema.
-    if (schema.iri == shaclSchemaUri)
-      for {
-        joinedTypes <- checkAndJoinTypes(value.graph.primaryTypes.map(_.value))
-        created     <- repo.create(id, schema, joinedTypes, value.source)
-      } yield created
-    else
-      for {
-        resolved    <- schemaContext(id, schema)
-        _           <- validate(resolved.schema, resolved.schemaImports, resolved.dataImports, value.graph)
-        joinedTypes <- checkAndJoinTypes(value.graph.primaryTypes.map(_.value))
-        created     <- repo.create(id, schema, joinedTypes, value.source)
-      } yield created
-  }
-
-  def fetch(id: ResId, schemaOpt: Option[Ref]): OptResource =
-    checkSchema(schemaOpt, repo.get(id))
-
-  def fetch(id: ResId, rev: Long, schemaOpt: Option[Ref]): OptResource =
-    checkSchema(schemaOpt, repo.get(id, rev))
-
-  def fetch(id: ResId, tag: String, schemaOpt: Option[Ref]): OptResource =
-    checkSchema(schemaOpt, repo.get(id, tag))
-
-  /**
-    * Updates an existing resource.
-    *
-    * @param id        the id of the resource
-    * @param rev       the last known revision of the resource
-    * @param schemaOpt optional schema reference that constrains the resource
-    * @param source    the new source representation in json-ld format
-    * @return either a rejection or the updated resource in the F context
-    */
-  def update(id: ResId, rev: Long, schemaOpt: Option[Ref], source: Json)(implicit identity: Identity): RejOrResource = {
-    def checkSchema(res: Resource): EitherT[F, Rejection, Unit] = schemaOpt match {
-      case Some(schema) if schema != res.schema => EitherT.leftT(NotFound(schema))
-      case _                                    => EitherT.rightT(())
-    }
-    // format: off
-    for {
-      resource    <- fetch(id, rev, None).toRight(NotFound(id.ref))
-      _           <- checkSchema(resource)
-      value       <- materialize(id, source)
-      graph        = value.graph
-      resolved    <- schemaContext(id, resource.schema)
-      _           <- validate(resolved.schema, resolved.schemaImports, resolved.dataImports, graph)
-      updated     <- repo.update(id, rev, graph.primaryTypes.map(_.value), source)
-    } yield updated
-  }
-  // format: on
-
-  def deprecate(id: ResId, rev: Long, schemaOpt: Option[Ref])(implicit identity: Identity): RejOrResource =
-    checkSchema(id, schemaOpt)(repo.deprecate(id, rev))
-
-  def tag(id: ResId, rev: Long, schemaOpt: Option[Ref], json: Json)(implicit identity: Identity): RejOrResource = {
-    val cursor = (json deepMerge Contexts.tagCtx).asGraph.cursor()
-    val result = for {
-      revValue <- cursor.downField(nxv.rev).focus.as[Long]
-      tagValue <- cursor.downField(nxv.tag).focus.as[String]
-    } yield tag(id, rev, schemaOpt, revValue, tagValue)
-    result match {
-      case Right(v) => v
-      case _        => EitherT.leftT(InvalidPayload(id.ref, "Both 'tag' and 'rev' fields must be present."))
-    }
-  }
-
-  def tag(id: ResId, rev: Long, schemaOpt: Option[Ref], targetRev: Long, tag: String)(
-      implicit identity: Identity): RejOrResource =
-    checkSchema(id, schemaOpt)(repo.tag(id, rev, targetRev, tag))
-
-  def attach[In](id: ResId, rev: Long, schemaOpt: Option[Ref], attach: BinaryDescription, source: In)(
-      implicit identity: Identity,
-      store: AttachmentStore[F, In, _]): RejOrResource =
-    checkSchema(id, schemaOpt)(repo.attach(id, rev, attach, source))
-
-  /**
-    * Removes an attachment from a resource.
-    *
-    * @param id        the id of the resource
-    * @param rev       the last known revision of the resource
-    * @param schemaOpt optional schema reference that constrains the resource
-    * @param filename  the attachment filename
-    * @return either a rejection or the new resource representation in the F context
-    */
-  def unattach(id: ResId, rev: Long, schemaOpt: Option[Ref], filename: String)(
-      implicit identity: Identity): RejOrResource =
-    checkSchema(id, schemaOpt)(repo.unattach(id, rev, filename))
-
-  /**
-    * Attempts to stream the resource's attachment identified by the argument id and the filename.
-    *
-    * @param id        the id of the resource.
-    * @param filename  the filename of the attachment
-    * @param schemaOpt optional schema reference that constrains the resource
-    * @tparam Out the type for the output streaming of the attachment binary
-    * @return the optional streamed attachment in the F context
-    */
-  def fetchAttachment[Out](id: ResId, schemaOpt: Option[Ref], filename: String)(
-      implicit store: AttachmentStore[F, _, Out]): OptionT[F, (BinaryAttributes, Out)] =
-    checkSchemaAtt(id, schemaOpt)(repo.getAttachment(id, filename))
-
-  /**
-    * Attempts to stream the resource's attachment identified by the argument id, the revision and the filename.
-    *
-    * @param id        the id of the resource.
-    * @param rev       the revision of the resource
-    * @param schemaOpt optional schema reference that constrains the resource
-    * @param filename  the filename of the attachment
-    * @tparam Out the type for the output streaming of the attachment binary
-    * @return the optional streamed attachment in the F context
-    */
-  def fetchAttachment[Out](id: ResId, rev: Long, schemaOpt: Option[Ref], filename: String)(
-      implicit store: AttachmentStore[F, _, Out]): OptionT[F, (BinaryAttributes, Out)] =
-    checkSchemaAtt(id, schemaOpt)(repo.getAttachment(id, rev, filename))
-
-  /**
-    * Attempts to stream the resource's attachment identified by the argument id, the tag and the filename. The
-    * tag is transformed into a revision value using the latest resource tag to revision mapping.
-    *
-    * @param id        the id of the resource.
-    * @param tag       the tag of the resource
-    * @param schemaOpt optional schema reference that constrains the resource
-    * @param filename  the filename of the attachment
-    * @tparam Out the type for the output streaming of the attachment binary
-    * @return the optional streamed attachment in the F context
-    */
-  def fetchAttachment[Out](id: ResId, tag: String, schemaOpt: Option[Ref], filename: String)(
-      implicit store: AttachmentStore[F, _, Out]): OptionT[F, (BinaryAttributes, Out)] =
-    checkSchemaAtt(id, schemaOpt)(repo.getAttachment(id, tag, filename))
-
-  /**
-    * Lists resources for the given project
-    *
-    * @param project        projects from which resources will be listed
-    * @param deprecated     deprecation status of the resources
-    * @param pagination     pagination options
-    * @param tc             typed HTTP client
-    * @return               search results in the F context
-    */
-  def list(project: ProjectRef, deprecated: Option[Boolean], pagination: Pagination)(
-      implicit tc: HttpClient[F, QueryResults[AbsoluteIri]]): F[QueryResults[AbsoluteIri]] = {
-    projects.views(project).flatMap { views =>
-      val elasticView = views.find {
-        case _: ElasticView => true
-        case _              => false
-      }
-      elasticView match {
-        case Some(view) =>
-          elastic.search(QueryBuilder.queryFor(deprecated), Set(ElasticIndexer.elasticIndex(view)))(pagination)
-        case None =>
-          F.pure(UnscoredQueryResults(0L, List.empty))
-      }
-    }
-  }
-
-  def list(project: ProjectRef, deprecated: Option[Boolean], schema: AbsoluteIri, pagination: Pagination)(
-      implicit tc: HttpClient[F, QueryResults[AbsoluteIri]]): F[QueryResults[AbsoluteIri]] = {
-    projects.views(project).flatMap { views =>
-      val elasticView = views.find {
-        case _: ElasticView => true
-        case _              => false
-      }
-      elasticView match {
-        case Some(view) =>
-          elastic.search(QueryBuilder.queryFor(deprecated, schema), Set(ElasticIndexer.elasticIndex(view)))(pagination)
-        case None =>
-          F.pure(UnscoredQueryResults(0L, List.empty))
-      }
-    }
-  }
-
-  private def materialize(projectRef: ProjectRef, source: Json): EitherT[F, Rejection, ResourceF.Value] = {
-
-    def flattenValue(refs: List[Ref], contextValue: Json): EitherT[F, Rejection, Json] =
-      (contextValue.asString, contextValue.asArray, contextValue.asObject) match {
-        case (Some(str), _, _) =>
-          val nextRef = Iri.absolute(str).toOption.map(Ref.apply)
-          for {
-            next  <- EitherT.fromOption[F](nextRef, IllegalContextValue(refs))
-            res   <- next.resolveOr(projectRef)(NotFound)
-            value <- flattenValue(next :: refs, res.value.contextValue)
-          } yield value
-        case (_, Some(arr), _) =>
-          import cats.implicits._
-          val jsons = arr
-            .traverse(j => flattenValue(refs, j).value)
-            .map(_.sequence)
-          EitherT(jsons).map(_.foldLeft(Json.obj())(_ deepMerge _))
-        case (_, _, Some(_)) => EitherT.rightT(contextValue)
-        case (_, _, _)       => EitherT.leftT(IllegalContextValue(refs))
-      }
-
-    def graphFor(flattenCtx: Json): Graph =
-      (source deepMerge Json.obj("@context" -> flattenCtx)).asGraph
-
-    flattenValue(Nil, source.contextValue)
-      .map(ctx => Value(source, ctx, graphFor(ctx)))
-  }
-
-  private def materialize(id: ResId, source: Json): EitherT[F, Rejection, ResourceF.Value] =
-    // format: off
-    for {
-      rawValue      <- materialize(id.parent, source)
-      value         <- checkOrAssignId(Left(id), rawValue)
-      (_, assigned)  = value
-    } yield assigned
-  // format: on
-
-  def materialize(resource: Resource): EitherT[F, Rejection, ResourceV] =
-    for {
-      value <- materialize(resource.id, resource.value)
-    } yield resource.map(_ => value)
-
-  def materializeWithMeta(resource: Resource): RejOrResourceV =
-    for {
-      resourceV <- materialize(resource)
-      graph = resourceV.value.graph
-      value = resourceV.value.copy(graph = graph ++ resourceV.metadata(_.iri) ++ resourceV.typeGraph)
-    } yield resourceV.map(_ => value)
-
-  /**
-    * Transitively imports resources referenced by the primary node of the resource through ''owl:imports'' if the
-    * resource has type ''owl:Ontology''.
-    *
-    * @param resource the resource for which imports are looked up
-    */
-  private def imports(resource: ResourceV): EitherT[F, Rejection, Set[ResourceV]] = {
-    import cats.implicits._
-    def canImport(id: AbsoluteIri, g: Graph): Boolean =
-      g.cursor(id)
-        .downField(rdf.tpe)
-        .values
-        .flatMap { vs =>
-          val set = vs.toSet
-          Some(set.contains(nxv.Schema) || set.contains(owl.Ontology))
-        }
-        .getOrElse(false)
-
-    def importsValues(id: AbsoluteIri, g: Graph): Set[Ref] =
-      if (canImport(id, g))
-        g.objects(IriNode(id), owl.imports).unorderedFoldMap {
-          case IriNode(iri) => Set(Ref(iri))
-          case _            => Set.empty
-        } else Set.empty
-
-    def lookup(current: Map[Ref, ResourceV], remaining: List[Ref]): EitherT[F, Rejection, Set[ResourceV]] = {
-      def load(ref: Ref): EitherT[F, Rejection, (Ref, ResourceV)] =
-        current
-          .find(_._1 == ref)
-          .map(tuple => EitherT.rightT[F, Rejection](tuple))
-          .getOrElse(ref.resolveOr(resource.id.parent)(NotFound).flatMap(r => materialize(r)).map(ref -> _))
-
-      if (remaining.isEmpty) EitherT.rightT(current.values.toSet)
-      else {
-        val batch: EitherT[F, Rejection, List[(Ref, ResourceV)]] =
-          remaining.traverse(ref => load(ref))
-
-        batch.flatMap { list =>
-          val nextRemaining: List[Ref] = list.flatMap {
-            case (ref, res) => importsValues(ref.iri, res.value.graph).toList
-          }
-          val nextCurrent: Map[Ref, ResourceV] = current ++ list.toMap
-          lookup(nextCurrent, nextRemaining)
-        }
-      }
-    }
-    lookup(Map.empty, importsValues(resource.id.value, resource.value.graph).toList)
-  }
-
-  private def checkSchemaAtt[Out](id: ResId, schemaOpt: Option[Ref])(
-      op: => OptionT[F, (BinaryAttributes, Out)]): OptionT[F, (BinaryAttributes, Out)] =
-    schemaOpt match {
-      case Some(_) => fetch(id, schemaOpt).flatMap(_ => op)
-      case _       => op
-    }
-
-  private def checkSchema(id: ResId, schemaOpt: Option[Ref])(op: => RejOrResource): RejOrResource =
-    schemaOpt match {
-      case Some(schema) => fetch(id, schemaOpt).toRight[Rejection](NotFound(schema)).flatMap(_ => op)
-      case _            => op
-    }
-
-  private def checkSchema(schemaOpt: Option[Ref], resource: OptResource): OptResource =
-    resource.flatMap { res =>
-      schemaOpt match {
-        case Some(schema) if schema != res.schema => OptionT.none[F, Resource]
-        case _                                    => resource
-      }
-    }
-
-  private def schemaContext(id: ResId, schema: Ref): EitherT[F, Rejection, SchemaContext] = {
-    def partition(set: Set[ResourceV]): (Set[ResourceV], Set[ResourceV]) =
-      set.partition(_.isSchema)
-    // format: off
-    for {
-      resolvedSchema                <- schema.resolveOr(id.parent)(NotFound)
-      materializedSchema            <- materialize(resolvedSchema)
-      importedResources             <- imports(materializedSchema)
-      (schemaImports, dataImports)  = partition(importedResources)
-    } yield SchemaContext(materializedSchema, dataImports, schemaImports)
-    // format: on
-  }
-
-  private def validate(schema: ResourceV,
-                       schemaImports: Set[ResourceV],
-                       dataImports: Set[ResourceV],
-                       data: Graph): EitherT[F, Rejection, Unit] = {
-    val resolvedSchema = schemaImports.foldLeft(schema.value.graph)(_ ++ _.value.graph)
-    val resolvedData   = dataImports.foldLeft(data)(_ ++ _.value.graph)
-    val report         = Validator.validate(resolvedSchema, resolvedData)
-    if (report.conforms) EitherT.rightT(())
-    else EitherT.leftT(InvalidResource(schema.id.ref, report))
-  }
-
-  private def checkOrAssignId(idOrGenInput: Either[ResId, (ProjectRef, AbsoluteIri)],
-                              value: ResourceF.Value): EitherT[F, Rejection, (ResId, ResourceF.Value)] = {
-
-    def replaceBNode(bnode: BNode, id: AbsoluteIri): ResourceF.Value =
-      value.copy(graph = value.graph.replaceNode(bnode, id))
-
-    def uuid(): String =
-      UUID.randomUUID().toString.toLowerCase
-
-    idOrGenInput match {
-      case Left(id) =>
-        if (value.graph.subjects().contains(id.value)) EitherT.rightT((id, value))
-        else {
-          val withIdOpt = value.graph.primaryBNode.map { bnode =>
-            (id, replaceBNode(bnode, id.value))
-          }
-          EitherT.fromOption(withIdOpt, IncorrectId(id.ref))
-        }
-      case Right((projectRef, base)) =>
-        val withIdOpt = value.graph.primaryNode map {
-          case IriNode(iri) => (Id(projectRef, iri), value)
-          case b: BNode =>
-            val id = Id(projectRef, base + uuid())
-            (id, replaceBNode(b, id.value))
-        }
-        EitherT.fromOption(withIdOpt, UnableToSelectResourceId)
-    }
-  }
-
-  private final implicit class RefSyntax(ref: Ref) {
-
-    def resolve(projectRef: ProjectRef): F[Option[Resource]] =
-      projects.resolution(projectRef)(self).resolve(ref)
-
-    def resolveOr(projectRef: ProjectRef)(f: Ref => Rejection): EitherT[F, Rejection, Resource] =
-      EitherT.fromOptionF(resolve(projectRef), f(ref))
-  }
-}
-object ResourcesImpl {
-  private[resources] final case class SchemaContext(schema: ResourceV,
-                                                    dataImports: Set[ResourceV],
-                                                    schemaImports: Set[ResourceV])
-
-}
