+++ src/main/scala/ch/epfl/bluebrain/nexus/kg/search/QueryBuilder.scala
@@ -28,22 +28,14 @@ object QueryBuilder {
   /**
     * Build Elastic search query from deprecation status and schema
     * @param deprecated optional deprecation status
-    * @param schema     schema to filter resources by
+    * @param schema     optional schema to filter resources by
     * @return           ElasticSearch query
     */
-  def queryFor(deprecated: Option[Boolean], schema: AbsoluteIri): Json = baseQuery(
-    Json.obj(
-      "term" -> Json.obj(
-        "_constrainedBy" -> Json.fromString(schema.show)
-      )
-    ) :: deprecatedTerm(deprecated)
-  )
-
-  /**
-    * Build Elastic search query from deprecation status
-    * @param deprecated optional deprecation status
-    * @return           ElasticSearch query
-    */
-  def queryFor(deprecated: Option[Boolean]): Json = baseQuery(deprecatedTerm(deprecated))
-
+  def queryFor(deprecated: Option[Boolean], schema: Option[AbsoluteIri] = None): Json =
+    baseQuery(schema match {
+      case Some(s) =>
+        Json.obj("term" -> Json.obj("_constrainedBy" -> Json.fromString(s.show))) :: deprecatedTerm(deprecated)
+      case None =>
+        deprecatedTerm(deprecated)
+    })
 }
